feat(products): update on producs; solar panels

+++ app/models.py
@@ -69,7 +69,7 @@ class ProductBrand(models.Model):
 
 
 class PanelSize(models.Model):
-    value = models.FloatField(primary_key=True)
+    value = models.IntegerField(primary_key=True)
 
 
 class SolarPanel(models.Model):
@@ -100,15 +100,18 @@ class DCCable(models.Model):
     length = models.FloatField()
     brand = models.ForeignKey(ProductBrand, on_delete=models.CASCADE)
 
+
 class Connector(models.Model):
     size = models.FloatField()
     length = models.FloatField()
     brand = models.ForeignKey(ProductBrand, on_delete=models.CASCADE)
 
+
 class Combiner(models.Model):
     size = models.FloatField()
     length = models.FloatField()
     brand = models.ForeignKey(ProductBrand, on_delete=models.CASCADE)
+
 
 class SpazrUserProduct(models.Model):
     user = models.ForeignKey(SpazrUser, on_delete=models.CASCADE)
